Revert "Merge pull request #6 from camiloaacevedo/shell-routes"
This reverts commit 404f852e03ff8db2539293b2ac54b66f75355612, reversing
changes made to 7b452caeab06dc75227ef2a3c2d9c1eb93c77552.
