+++ lib/presentation/delegates/search_movie_delegate.dart
@@ -15,16 +15,13 @@ class SearchMovieDelegate extends SearchDelegate<Movie?>{
   List<Movie> initialMovies;
 
   StreamController<List<Movie>> debouncedMovies = StreamController.broadcast();
-  StreamController<bool> isLoadingStream = StreamController.broadcast();
-  
-  
   Timer? _debounceTimer;
 
   SearchMovieDelegate({
     required this.searchMovies,
     required this.initialMovies,
   }):super(
-    searchFieldLabel: 'Search movies',
+    searchFieldLabel: 'Buscar películas',
     // textInputAction: TextInputAction.done
   );
 
@@ -33,8 +30,6 @@ class SearchMovieDelegate extends SearchDelegate<Movie?>{
   }
 
   void _onQueryChanged( String query ) {
-    isLoadingStream.add(true);
-
     if (_debounceTimer?.isActive ?? false ) _debounceTimer!.cancel();
 
     _debounceTimer = Timer(const Duration( milliseconds: 500), () async {
@@ -46,7 +41,6 @@ class SearchMovieDelegate extends SearchDelegate<Movie?>{
       final movies = await searchMovies( query );
       initialMovies = movies;
       debouncedMovies.add(movies);
-      isLoadingStream.add(false);
 
     });
 
@@ -82,33 +76,15 @@ class SearchMovieDelegate extends SearchDelegate<Movie?>{
   List<Widget>? buildActions(BuildContext context) {
 
     return [
+      
+        FadeIn(
+          animate: query.isNotEmpty,
+          child: IconButton(
+            onPressed: () => query = '', 
+            icon: const Icon( Icons.clear )
+          ),
+        ),
 
-      StreamBuilder(
-        initialData: false,
-        stream: isLoadingStream.stream,
-        builder: (context, snapshot) {
-            if ( snapshot.data ?? false ) {
-              return SpinPerfect(
-                  duration: const Duration(seconds: 20),
-                  spins: 10,
-                  infinite: true,
-                  child: IconButton(
-                    onPressed: () => query = '', 
-                    icon: const Icon( Icons.refresh_rounded )
-                  ),
-                );
-            }
-
-            return FadeIn(
-              animate: query.isNotEmpty,
-              child: IconButton(
-                onPressed: () => query = '', 
-                icon: const Icon( Icons.clear )
-              ),
-            );
-
-        },
-      ),
 
     ];
   }
