+++ lib/presentation/screens/movies/home_screen.dart
@@ -1,5 +1,7 @@
+import 'package:cinemapedia/presentation/providers/providers.dart';
 import 'package:flutter/material.dart';
 import 'package:cinemapedia/presentation/widgets/widgets.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 
 
@@ -7,19 +9,109 @@ class HomeScreen extends StatelessWidget {
 
   static const name = 'home-screen';
 
-  final Widget childView;
-
-
-  const HomeScreen({
-    super.key, 
-    required this.childView
-  });
+  const HomeScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: childView,
-      bottomNavigationBar: const CustomBottomNavigation(),
+    return const Scaffold(
+      body: _HomeView(),
+      bottomNavigationBar: CustomBottomNavigation(),
+    );
+  }
+}
+
+class _HomeView extends ConsumerStatefulWidget {
+  const _HomeView();
+
+  @override
+  _HomeViewState createState() => _HomeViewState();
+}
+
+class _HomeViewState extends ConsumerState<_HomeView> {
+
+  @override
+  void initState() {
+    super.initState();
+
+    ref.read( nowPlayingMoviesProvider.notifier ).loadNextPage();
+    ref.read( popularMoviesProvider.notifier ).loadNextPage();
+    ref.read( topRatedMoviesProvider.notifier ).loadNextPage();
+    ref.read( upcomingMoviesProvider.notifier ).loadNextPage();
+  }
+
+
+  @override
+  Widget build(BuildContext context) {
+
+    final initialLoading = ref.watch(initialLoadingProvider);
+    if ( initialLoading ) return const FullScreenLoader();
+
+    final slideShowMovies = ref.watch( moviesSlideshowProvider );
+    final nowPlayingMovies = ref.watch( nowPlayingMoviesProvider );
+    final popularMovies = ref.watch( popularMoviesProvider );
+    final topRatedMovies = ref.watch( topRatedMoviesProvider );
+    final upcomingMovies = ref.watch( upcomingMoviesProvider );
+
+    return CustomScrollView(
+      slivers: [
+
+        const SliverAppBar(
+          floating: true,
+          flexibleSpace: FlexibleSpaceBar(
+            title: CustomAppbar(),
+            titlePadding: EdgeInsets.zero,
+          ),
+        ),
+
+
+        SliverList(delegate: SliverChildBuilderDelegate(
+          (context, index) {
+            return Column(
+                children: [
+
+                  // const CustomAppbar(),
+
+                  MoviesSlideshow(movies: slideShowMovies ),
+
+                  MovieHorizontalListview(
+                    movies: nowPlayingMovies,
+                    title: 'In theaters',
+                    subTitle: 'Monday 20',
+                    loadNextPage: () =>ref.read(nowPlayingMoviesProvider.notifier).loadNextPage(),
+
+                  ),
+
+                  MovieHorizontalListview(
+                    movies: upcomingMovies,
+                    title: 'Soon',
+                    subTitle: 'This month',
+                    loadNextPage: () =>ref.read(upcomingMoviesProvider.notifier).loadNextPage(),
+                  ),
+
+                  MovieHorizontalListview(
+                    movies: popularMovies,
+                    title: 'Popular',
+                    // subTitle: '',
+                    loadNextPage: () =>ref.read(popularMoviesProvider.notifier).loadNextPage(),
+                  ),
+
+                  MovieHorizontalListview(
+                    movies: topRatedMovies,
+                    title: 'Best rated',
+                    subTitle: 'Since always',
+                    loadNextPage: () =>ref.read(topRatedMoviesProvider.notifier).loadNextPage(), 
+                  ),
+
+                  const SizedBox( height: 10 ),
+
+
+              ],
+            );
+          },
+          childCount: 1
+        )),
+
+      ]
     );
   }
 }
