+++ lib/presentation/views/home_views/favorites_view.dart
@@ -1,36 +1,20 @@
-import 'package:cinemapedia/presentation/widgets/movies/movie_masonry.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:cinemapedia/presentation/providers/providers.dart';
 
 
-class FavoritesView extends ConsumerStatefulWidget {
+
+class FavoritesView extends StatelessWidget {
   const FavoritesView({super.key});
 
   @override
-  FavoritesViewState createState() => FavoritesViewState();
-}
-
-class FavoritesViewState extends ConsumerState<FavoritesView> {
-
-  @override
-  void initState() {
-    super.initState();
-
-    ref.read(favoriteMoviesProvider.notifier)
-      .loadNextPage();
-
-  }
-
-
-  @override
   Widget build(BuildContext context) {
-
-    final favoriteMovies = ref.watch(favoriteMoviesProvider).values.toList();
-
-
     return Scaffold(
-      body: MovieMasonry(movies: favoriteMovies)
+      appBar: AppBar(
+        title: Text('Favorites View'),
+        centerTitle: true,
+      ),
+      body: Center(
+        child: const Text('Favorites'),
+      ),
     );
   }
 }
