--- lib/presentation/views/views.dart
@@ -1,4 +1,0 @@
-
-
-export 'home_views/favorites_view.dart';
-export 'home_views/home_view.dart';
